fixed docs and moved galaxy
