+++ .github/workflows/pages.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Build Documentation
         run: |
           cd docs
-          sphinx-apidoc -o source ../src/galaxy
+          sphinx-apidoc -o source ../galaxy
           sphinx-apidoc -o source ../API
           make html
