+++ docs/CONFIG_DOC.md
@@ -11,7 +11,7 @@ Before getting started on config Make sure you have [Postgres](https://www.postg
 ### 2. Prepare your OSM Snapshot Data
 Initialize rawdata from [here](https://github.com/hotosm/underpass/tree/master/raw) OR Create database "raw" in your local postgres and insert sample dump from
 ```
-/tests/src/fixtures/raw_data.sql
+/tests/fixtures/raw_data.sql
 ```
 
 ```
@@ -72,7 +72,7 @@ Considering You have PSQL-POSTGIS setup  with user **postgres** host **localhost
   export PGPASSWORD='admin';
   psql -U postgres -h localhost -p 5432 -c "CREATE DATABASE raw;"
 
-  cd tests/src/fixtures/
+  cd tests/fixtures/
   psql -U postgres -h localhost -p 5432 raw  < raw_data.sql
 ```
