+++ docs/source/modules.rst
@@ -1,4 +1,4 @@
-Galaxy
+galaxy
 ======
 
 .. toctree::
