--- src/galaxy/__init__.py
@@ -1,9 +1,0 @@
-from .app import Database, RawData
-from .config import config
-
-
-__all__ = [
-    'Database',
-    'config',
-    'RawData'
-]
